+++ src/periph/uarte/lib.rs
@@ -1,0 +1,688 @@
+//! Universal Asynchronous Receiver/Transmitter.
+
+#![feature(proc_macro_hygiene)]
+#![deny(elided_lifetimes_in_paths)]
+#![warn(missing_docs)]
+#![warn(clippy::pedantic)]
+#![allow(clippy::must_use_candidate, clippy::type_repetition_in_bounds)]
+#![no_std]
+
+use drone_core::periph;
+use drone_cortex_m::reg::marker::*;
+
+periph! {
+    /// Generic UARTE peripheral variant.
+    pub trait UarteMap {}
+
+    /// Generic UARTE peripheral.
+    pub struct UartePeriph;
+
+    UARTE {
+        TASKS_STARTRX {
+            0x20 WoReg;
+            TASKS_STARTRX { WoWoRegFieldBit }
+        }
+        TASKS_STOPRX {
+            0x20 WoReg;
+            TASKS_STOPRX { WoWoRegFieldBit }
+        }
+        TASKS_STARTTX {
+            0x20 WoReg;
+            TASKS_STARTTX { WoWoRegFieldBit }
+        }
+        TASKS_STOPTX {
+            0x20 WoReg;
+            TASKS_STOPTX { WoWoRegFieldBit }
+        }
+        TASKS_FLUSHRX {
+            0x20 WoReg;
+            TASKS_FLUSHRX { WoWoRegFieldBit }
+        }
+        SUBSCRIBE_STARTRX {
+            0x20 RwReg;
+            CHIDX { RwRwRegFieldBits }
+            EN { RwRwRegFieldBit }
+        }
+        SUBSCRIBE_STOPRX {
+            0x20 RwReg;
+            CHIDX { RwRwRegFieldBits }
+            EN { RwRwRegFieldBit }
+        }
+        SUBSCRIBE_STARTTX {
+            0x20 RwReg;
+            CHIDX { RwRwRegFieldBits }
+            EN { RwRwRegFieldBit }
+        }
+        SUBSCRIBE_STOPTX {
+            0x20 RwReg;
+            CHIDX { RwRwRegFieldBits }
+            EN { RwRwRegFieldBit }
+        }
+        SUBSCRIBE_FLUSHRX {
+            0x20 RwReg;
+            CHIDX { RwRwRegFieldBits }
+            EN { RwRwRegFieldBit }
+        }
+        EVENTS_CTS {
+            0x20 RwReg;
+            EVENTS_CTS { RwRwRegFieldBit }
+        }
+        EVENTS_NCTS {
+            0x20 RwReg;
+            EVENTS_NCTS { RwRwRegFieldBit }
+        }
+        EVENTS_RXDRDY {
+            0x20 RwReg;
+            EVENTS_RXDRDY { RwRwRegFieldBit }
+        }
+        EVENTS_ENDRX {
+            0x20 RwReg;
+            EVENTS_ENDRX { RwRwRegFieldBit }
+        }
+        EVENTS_TXDRDY {
+            0x20 RwReg;
+            EVENTS_TXDRDY { RwRwRegFieldBit }
+        }
+        EVENTS_ENDTX {
+            0x20 RwReg;
+            EVENTS_ENDTX { RwRwRegFieldBit }
+        }
+        EVENTS_ERROR {
+            0x20 RwReg;
+            EVENTS_ERROR { RwRwRegFieldBit }
+        }
+        EVENTS_RXTO {
+            0x20 RwReg;
+            EVENTS_RXTO { RwRwRegFieldBit }
+        }
+        EVENTS_RXSTARTED {
+            0x20 RwReg;
+            EVENTS_RXSTARTED { RwRwRegFieldBit }
+        }
+        EVENTS_TXSTARTED {
+            0x20 RwReg;
+            EVENTS_TXSTARTED { RwRwRegFieldBit }
+        }
+        EVENTS_TXSTOPPED {
+            0x20 RwReg;
+            EVENTS_TXSTOPPED { RwRwRegFieldBit }
+        }
+        PUBLISH_CTS {
+            0x20 RwReg;
+            CHIDX { RwRwRegFieldBits }
+            EN { RwRwRegFieldBit }
+        }
+        PUBLISH_NCTS {
+            0x20 RwReg;
+            CHIDX { RwRwRegFieldBits }
+            EN { RwRwRegFieldBit }
+        }
+        PUBLISH_RXDRDY {
+            0x20 RwReg;
+            CHIDX { RwRwRegFieldBits }
+            EN { RwRwRegFieldBit }
+        }
+        PUBLISH_ENDRX {
+            0x20 RwReg;
+            CHIDX { RwRwRegFieldBits }
+            EN { RwRwRegFieldBit }
+        }
+        PUBLISH_TXDRDY {
+            0x20 RwReg;
+            CHIDX { RwRwRegFieldBits }
+            EN { RwRwRegFieldBit }
+        }
+        PUBLISH_ENDTX {
+            0x20 RwReg;
+            CHIDX { RwRwRegFieldBits }
+            EN { RwRwRegFieldBit }
+        }
+        PUBLISH_ERROR {
+            0x20 RwReg;
+            CHIDX { RwRwRegFieldBits }
+            EN { RwRwRegFieldBit }
+        }
+        PUBLISH_RXTO {
+            0x20 RwReg;
+            CHIDX { RwRwRegFieldBits }
+            EN { RwRwRegFieldBit }
+        }
+        PUBLISH_RXSTARTED {
+            0x20 RwReg;
+            CHIDX { RwRwRegFieldBits }
+            EN { RwRwRegFieldBit }
+        }
+        PUBLISH_TXSTARTED {
+            0x20 RwReg;
+            CHIDX { RwRwRegFieldBits }
+            EN { RwRwRegFieldBit }
+        }
+        PUBLISH_TXSTOPPED {
+            0x20 RwReg;
+            CHIDX { RwRwRegFieldBits }
+            EN { RwRwRegFieldBit }
+        }
+        SHORTS {
+            0x20 RwReg;
+            ENDRX_STARTRX { RwRwRegFieldBit }
+            ENDRX_STOPRX { RwRwRegFieldBit }
+        }
+        INTEN {
+            0x20 RwReg;
+            CTS { RwRwRegFieldBit }
+            NCTS { RwRwRegFieldBit }
+            RXDRDY { RwRwRegFieldBit }
+            ENDRX { RwRwRegFieldBit }
+            TXDRDY { RwRwRegFieldBit }
+            ENDTX { RwRwRegFieldBit }
+            ERROR { RwRwRegFieldBit }
+            RXTO { RwRwRegFieldBit }
+            RXSTARTED { RwRwRegFieldBit }
+            TXSTARTED { RwRwRegFieldBit }
+            TXSTOPPED { RwRwRegFieldBit }
+        }
+        INTENSET {
+            0x20 RwReg;
+            CTS { RwRwRegFieldBit }
+            NCTS { RwRwRegFieldBit }
+            RXDRDY { RwRwRegFieldBit }
+            ENDRX { RwRwRegFieldBit }
+            TXDRDY { RwRwRegFieldBit }
+            ENDTX { RwRwRegFieldBit }
+            ERROR { RwRwRegFieldBit }
+            RXTO { RwRwRegFieldBit }
+            RXSTARTED { RwRwRegFieldBit }
+            TXSTARTED { RwRwRegFieldBit }
+            TXSTOPPED { RwRwRegFieldBit }
+        }
+        INTENCLR {
+            0x20 RwReg;
+            CTS { RwRwRegFieldBit }
+            NCTS { RwRwRegFieldBit }
+            RXDRDY { RwRwRegFieldBit }
+            ENDRX { RwRwRegFieldBit }
+            TXDRDY { RwRwRegFieldBit }
+            ENDTX { RwRwRegFieldBit }
+            ERROR { RwRwRegFieldBit }
+            RXTO { RwRwRegFieldBit }
+            RXSTARTED { RwRwRegFieldBit }
+            TXSTARTED { RwRwRegFieldBit }
+            TXSTOPPED { RwRwRegFieldBit }
+        }
+        ERRORSRC {
+            0x20 RwReg;
+            OVERRUN { RwRwRegFieldBit }
+            PARITY { RwRwRegFieldBit }
+            FRAMING { RwRwRegFieldBit }
+            BREAK { RwRwRegFieldBit }
+        }
+        ENABLE {
+            0x20 RwReg;
+            ENABLE { RwRwRegFieldBits }
+        }
+        PSEL_RTS {
+            0x20 RwReg;
+            PIN { RwRwRegFieldBits }
+            CONNECT { RwRwRegFieldBit }
+        }
+        PSEL_TXD {
+            0x20 RwReg;
+            PIN { RwRwRegFieldBits }
+            CONNECT { RwRwRegFieldBit }
+        }
+        PSEL_CTS {
+            0x20 RwReg;
+            PIN { RwRwRegFieldBits }
+            CONNECT { RwRwRegFieldBit }
+        }
+        PSEL_RXD {
+            0x20 RwReg;
+            PIN { RwRwRegFieldBits }
+            CONNECT { RwRwRegFieldBit }
+        }
+        BAUDRATE {
+            0x20 RwReg;
+            BAUDRATE { RwRwRegFieldBits }
+        }
+        RXD_PTR {
+            0x20 RwReg;
+            PTR { RwRwRegFieldBits }
+        }
+        RXD_MAXCNT {
+            0x20 RwReg;
+            MAXCNT { RwRwRegFieldBits }
+        }
+        RXD_AMOUNT {
+            0x20 RoReg;
+            AMOUNT { RoRoRegFieldBits }
+        }
+        TXD_PTR {
+            0x20 RwReg;
+            PTR { RwRwRegFieldBits }
+        }
+        TXD_MAXCNT {
+            0x20 RwReg;
+            MAXCNT { RwRwRegFieldBits }
+        }
+        TXD_AMOUNT {
+            0x20 RoReg;
+            AMOUNT { RoRoRegFieldBits }
+        }
+        CONFIG {
+            0x20 RwReg;
+            HWFC { RwRwRegFieldBit }
+            PARITY { RwRwRegFieldBits }
+            STOP { RwRwRegFieldBit }
+        }
+    }
+}
+
+#[allow(unused_macros)]
+macro_rules! map_uarte_ns_s {
+    (
+        $uarte_ns_macro_doc:expr,
+        $uarte_s_macro_doc:expr,
+        $uarte_ns_macro:ident,
+        $uarte_s_macro:ident,
+        $uarte_ns_ty_doc:expr,
+        $uarte_s_ty_doc:expr,
+        $uarte_ns_ty:ident,
+        $uarte_s_ty:ident,
+        $uarte_ns:ident,
+        $uarte_s:ident,
+        $spim_ns:ident,
+        $spim_s:ident,
+        $spis_ns:ident,
+        $spis_s:ident,
+        $twim_ns:ident,
+        $twim_s:ident,
+    ) => {
+        map_uarte!(
+            $uarte_ns_macro_doc,
+            $uarte_ns_macro,
+            $uarte_ns_ty_doc,
+            $uarte_ns_ty,
+            $uarte_ns,
+            $spim_ns,
+            $spis_ns,
+            $twim_ns,
+        );
+        map_uarte!(
+            $uarte_s_macro_doc,
+            $uarte_s_macro,
+            $uarte_s_ty_doc,
+            $uarte_s_ty,
+            $uarte_s,
+            $spim_s,
+            $spis_s,
+            $twim_s,
+        );
+    };
+}
+
+#[allow(unused_macros)]
+macro_rules! map_uarte {
+    (
+        $uarte_macro_doc:expr,
+        $uarte_macro:ident,
+        $uarte_ty_doc:expr,
+        $uarte_ty:ident,
+        $uarte:ident,
+        $spim:ident,
+        $spis:ident,
+        $twim:ident,
+    ) => {
+        periph::map! {
+            #[doc = $uarte_macro_doc]
+            pub macro $uarte_macro;
+
+            #[doc = $uarte_ty_doc]
+            pub struct $uarte_ty;
+
+            impl UarteMap for $uarte_ty {}
+
+            drone_nrf_map_pieces::reg;
+            crate;
+
+            UARTE {
+                $uarte;
+                TASKS_STARTRX {
+                    TASKS_STARTRX($twim TASKS_STARTRX);
+                    TASKS_STARTRX { TASKS_STARTRX }
+                }
+                TASKS_STOPRX {
+                    TASKS_STOPRX;
+                    TASKS_STOPRX { TASKS_STOPRX }
+                }
+                TASKS_STARTTX {
+                    TASKS_STARTTX($twim TASKS_STARTTX);
+                    TASKS_STARTTX { TASKS_STARTTX }
+                }
+                TASKS_STOPTX {
+                    TASKS_STOPTX;
+                    TASKS_STOPTX { TASKS_STOPTX }
+                }
+                TASKS_FLUSHRX {
+                    TASKS_FLUSHRX;
+                    TASKS_FLUSHRX { TASKS_FLUSHRX }
+                }
+                SUBSCRIBE_STARTRX {
+                    SUBSCRIBE_STARTRX($twim SUBSCRIBE_STARTRX);
+                    CHIDX { CHIDX }
+                    EN { EN }
+                }
+                SUBSCRIBE_STOPRX {
+                    SUBSCRIBE_STOPRX;
+                    CHIDX { CHIDX }
+                    EN { EN }
+                }
+                SUBSCRIBE_STARTTX {
+                    SUBSCRIBE_STARTTX($twim SUBSCRIBE_STARTTX);
+                    CHIDX { CHIDX }
+                    EN { EN }
+                }
+                SUBSCRIBE_STOPTX {
+                    SUBSCRIBE_STOPTX;
+                    CHIDX { CHIDX }
+                    EN { EN }
+                }
+                SUBSCRIBE_FLUSHRX {
+                    SUBSCRIBE_FLUSHRX;
+                    CHIDX { CHIDX }
+                    EN { EN }
+                }
+                EVENTS_CTS {
+                    EVENTS_CTS;
+                    EVENTS_CTS { EVENTS_CTS }
+                }
+                EVENTS_NCTS {
+                    EVENTS_NCTS($spim EVENTS_STOPPED);
+                    EVENTS_NCTS { EVENTS_NCTS }
+                }
+                EVENTS_RXDRDY {
+                    EVENTS_RXDRDY;
+                    EVENTS_RXDRDY { EVENTS_RXDRDY }
+                }
+                EVENTS_ENDRX {
+                    EVENTS_ENDRX($spim EVENTS_ENDRX);
+                    EVENTS_ENDRX { EVENTS_ENDRX }
+                }
+                EVENTS_TXDRDY {
+                    EVENTS_TXDRDY;
+                    EVENTS_TXDRDY { EVENTS_TXDRDY }
+                }
+                EVENTS_ENDTX {
+                    EVENTS_ENDTX($spim EVENTS_ENDTX);
+                    EVENTS_ENDTX { EVENTS_ENDTX }
+                }
+                EVENTS_ERROR {
+                    EVENTS_ERROR($twim EVENTS_ERROR);
+                    EVENTS_ERROR { EVENTS_ERROR }
+                }
+                EVENTS_RXTO {
+                    EVENTS_RXTO;
+                    EVENTS_RXTO { EVENTS_RXTO }
+                }
+                EVENTS_RXSTARTED {
+                    EVENTS_RXSTARTED($spim EVENTS_STARTED);
+                    EVENTS_RXSTARTED { EVENTS_RXSTARTED }
+                }
+                EVENTS_TXSTARTED {
+                    EVENTS_TXSTARTED($twim EVENTS_TXSTARTED);
+                    EVENTS_TXSTARTED { EVENTS_TXSTARTED }
+                }
+                EVENTS_TXSTOPPED {
+                    EVENTS_TXSTOPPED;
+                    EVENTS_TXSTOPPED { EVENTS_TXSTOPPED }
+                }
+                PUBLISH_CTS {
+                    PUBLISH_CTS;
+                    CHIDX { CHIDX }
+                    EN { EN }
+                }
+                PUBLISH_NCTS {
+                    PUBLISH_NCTS($spim PUBLISH_STOPPED);
+                    CHIDX { CHIDX }
+                    EN { EN }
+                }
+                PUBLISH_RXDRDY {
+                    PUBLISH_RXDRDY;
+                    CHIDX { CHIDX }
+                    EN { EN }
+                }
+                PUBLISH_ENDRX {
+                    PUBLISH_ENDRX($spim PUBLISH_ENDRX);
+                    CHIDX { CHIDX }
+                    EN { EN }
+                }
+                PUBLISH_TXDRDY {
+                    PUBLISH_TXDRDY;
+                    CHIDX { CHIDX }
+                    EN { EN }
+                }
+                PUBLISH_ENDTX {
+                    PUBLISH_ENDTX($spim PUBLISH_ENDTX);
+                    CHIDX { CHIDX }
+                    EN { EN }
+                }
+                PUBLISH_ERROR {
+                    PUBLISH_ERROR($twim PUBLISH_ERROR);
+                    CHIDX { CHIDX }
+                    EN { EN }
+                }
+                PUBLISH_RXTO {
+                    PUBLISH_RXTO;
+                    CHIDX { CHIDX }
+                    EN { EN }
+                }
+                PUBLISH_RXSTARTED {
+                    PUBLISH_RXSTARTED($spim PUBLISH_STARTED);
+                    CHIDX { CHIDX }
+                    EN { EN }
+                }
+                PUBLISH_TXSTARTED {
+                    PUBLISH_TXSTARTED($twim PUBLISH_TXSTARTED);
+                    CHIDX { CHIDX }
+                    EN { EN }
+                }
+                PUBLISH_TXSTOPPED {
+                    PUBLISH_TXSTOPPED;
+                    CHIDX { CHIDX }
+                    EN { EN }
+                }
+                SHORTS {
+                    SHORTS($spim SHORTS);
+                    ENDRX_STARTRX { ENDRX_STARTRX }
+                    ENDRX_STOPRX { ENDRX_STOPRX }
+                }
+                INTEN {
+                    INTEN($twim INTEN);
+                    CTS { CTS }
+                    NCTS { NCTS }
+                    RXDRDY { RXDRDY }
+                    ENDRX { ENDRX }
+                    TXDRDY { TXDRDY }
+                    ENDTX { ENDTX }
+                    ERROR { ERROR }
+                    RXTO { RXTO }
+                    RXSTARTED { RXSTARTED }
+                    TXSTARTED { TXSTARTED }
+                    TXSTOPPED { TXSTOPPED }
+                }
+                INTENSET {
+                    INTENSET($spim INTENSET);
+                    CTS { CTS }
+                    NCTS { NCTS }
+                    RXDRDY { RXDRDY }
+                    ENDRX { ENDRX }
+                    TXDRDY { TXDRDY }
+                    ENDTX { ENDTX }
+                    ERROR { ERROR }
+                    RXTO { RXTO }
+                    RXSTARTED { RXSTARTED }
+                    TXSTARTED { TXSTARTED }
+                    TXSTOPPED { TXSTOPPED }
+                }
+                INTENCLR {
+                    INTENCLR($spim INTENCLR);
+                    CTS { CTS }
+                    NCTS { NCTS }
+                    RXDRDY { RXDRDY }
+                    ENDRX { ENDRX }
+                    TXDRDY { TXDRDY }
+                    ENDTX { ENDTX }
+                    ERROR { ERROR }
+                    RXTO { RXTO }
+                    RXSTARTED { RXSTARTED }
+                    TXSTARTED { TXSTARTED }
+                    TXSTOPPED { TXSTOPPED }
+                }
+                ERRORSRC {
+                    ERRORSRC;
+                    OVERRUN { OVERRUN }
+                    PARITY { PARITY }
+                    FRAMING { FRAMING }
+                    BREAK { BREAK }
+                }
+                ENABLE {
+                    ENABLE($spim ENABLE);
+                    ENABLE { ENABLE }
+                }
+                PSEL_RTS {
+                    PSEL_RTS($spim PSEL_SCK);
+                    PIN { PIN }
+                    CONNECT { CONNECT }
+                }
+                PSEL_TXD {
+                    PSEL_TXD($spim PSEL_MOSI);
+                    PIN { PIN }
+                    CONNECT { CONNECT }
+                }
+                PSEL_CTS {
+                    PSEL_CTS($spim PSEL_MISO);
+                    PIN { PIN }
+                    CONNECT { CONNECT }
+                }
+                PSEL_RXD {
+                    PSEL_RXD($spis PSEL_CSN);
+                    PIN { PIN }
+                    CONNECT { CONNECT }
+                }
+                BAUDRATE {
+                    BAUDRATE($spim FREQUENCY);
+                    BAUDRATE { BAUDRATE }
+                }
+                RXD_PTR {
+                    RXD_PTR($spim RXD_PTR);
+                    PTR { PTR }
+                }
+                RXD_MAXCNT {
+                    RXD_MAXCNT($spim RXD_MAXCNT);
+                    MAXCNT { MAXCNT }
+                }
+                RXD_AMOUNT {
+                    RXD_AMOUNT($spim RXD_AMOUNT);
+                    AMOUNT { AMOUNT }
+                }
+                TXD_PTR {
+                    TXD_PTR($spim TXD_PTR);
+                    PTR { PTR }
+                }
+                TXD_MAXCNT {
+                    TXD_MAXCNT($spim TXD_MAXCNT);
+                    MAXCNT { MAXCNT }
+                }
+                TXD_AMOUNT {
+                    TXD_AMOUNT($spim TXD_AMOUNT);
+                    AMOUNT { AMOUNT }
+                }
+                CONFIG {
+                    CONFIG;
+                    HWFC { HWFC }
+                    PARITY { PARITY }
+                    STOP { STOP }
+                }
+            }
+        }
+    };
+}
+
+#[cfg(nrf_mcu = "nrf9160")]
+map_uarte_ns_s! {
+    "Extracts UARTE0_NS register tokens.",
+    "Extracts UARTE0_S register tokens.",
+    periph_uarte0_ns,
+    periph_uarte0_s,
+    "UARTE0_NS peripheral variant.",
+    "UARTE0_S peripheral variant.",
+    Uarte0Ns,
+    Uarte0S,
+    UARTE0_NS,
+    UARTE0_S,
+    SPIM0_NS,
+    SPIM0_S,
+    SPIS0_NS,
+    SPIS0_S,
+    TWIM0_NS,
+    TWIM0_S,
+}
+
+#[cfg(nrf_mcu = "nrf9160")]
+map_uarte_ns_s! {
+    "Extracts UARTE1_NS register tokens.",
+    "Extracts UARTE1_S register tokens.",
+    periph_uarte1_ns,
+    periph_uarte1_s,
+    "UARTE1_NS peripheral variant.",
+    "UARTE1_S peripheral variant.",
+    Uarte1Ns,
+    Uarte1S,
+    UARTE1_NS,
+    UARTE1_S,
+    SPIM1_NS,
+    SPIM1_S,
+    SPIS1_NS,
+    SPIS1_S,
+    TWIM1_NS,
+    TWIM1_S,
+}
+
+#[cfg(nrf_mcu = "nrf9160")]
+map_uarte_ns_s! {
+    "Extracts UARTE2_NS register tokens.",
+    "Extracts UARTE2_S register tokens.",
+    periph_uarte2_ns,
+    periph_uarte2_s,
+    "UARTE2_NS peripheral variant.",
+    "UARTE2_S peripheral variant.",
+    Uarte2Ns,
+    Uarte2S,
+    UARTE2_NS,
+    UARTE2_S,
+    SPIM2_NS,
+    SPIM2_S,
+    SPIS2_NS,
+    SPIS2_S,
+    TWIM2_NS,
+    TWIM2_S,
+}
+
+#[cfg(nrf_mcu = "nrf9160")]
+map_uarte_ns_s! {
+    "Extracts UARTE3_NS register tokens.",
+    "Extracts UARTE3_S register tokens.",
+    periph_uarte3_ns,
+    periph_uarte3_s,
+    "UARTE3_NS peripheral variant.",
+    "UARTE3_S peripheral variant.",
+    Uarte3Ns,
+    Uarte3S,
+    UARTE3_NS,
+    UARTE3_S,
+    SPIM3_NS,
+    SPIM3_S,
+    SPIS3_NS,
+    SPIS3_S,
+    TWIM3_NS,
+    TWIM3_S,
+}